requirements.txt
# Core AI & Logic
openai>=1.0.0            # OpenAI API for Copilot-style AI
pydantic>=2.12.0         # Latest Pydantic V2 for better performance
python-dotenv>=1.2.0     # Modern .env handling

# Web & API
fastapi>=0.128.0         # Modern FastAPI with lifespan support
uvicorn>=0.30.0          # Stable ASGI server
python-multipart>=0.0.20 # Required for file/form handling in FastAPI

# Extraction & Scraping
beautifulsoup4>=4.12.0   # Standard for HTML parsing
requests>=2.32.0         # Standard for sync URL fetching
pypdf>=6.6.0             # significantly faster PDF parsing (replaces older 4.x)

# Optional but Recommended
httpx>=0.28.0            # If you plan to use async URL fetching